💡 [ADD]: 3372. Maximize the Number of Target Nodes After Connecting Trees I
Implemented solution for LeetCode 3372. The problem requires maximizing target nodes for each node in Tree1 by connecting it to an optimal node in Tree2. The core observation is that for a node 'i' in Tree1 connected to 'j' in Tree2, the total target nodes are (nodes in T1 within distance 'k' of 'i') + (nodes in T2 within distance 'k-1' of 'j'). The solution calculates these counts for all nodes in both trees using N (or M) Breadth-First Searches. This approach has a time complexity of O(N^2 + M^2) and space complexity of O(N + M), which is efficient enough for the given constraints (N, M <= 1000).